@@ -1,14 +1,14 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 27
+    compileSdkVersion 30
     defaultConfig {
         applicationId "org.dreamexposure.startapped"
-        minSdkVersion 21
-        targetSdkVersion 27
+        minSdkVersion 30
+        targetSdkVersion 30
         versionCode 1
         versionName "1.0.0"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
     }
     buildTypes {
         release {
@@ -20,36 +20,38 @@ android {
         sourceCompatibility = '1.8'
         targetCompatibility = '1.8'
     }
+    buildToolsVersion '30.0.2'
 }
 
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation 'com.android.support:appcompat-v7:27.1.1'
-    implementation 'com.android.support.constraint:constraint-layout:1.1.3'
-    implementation 'com.android.support:design:27.1.1'
-    implementation 'com.android.support:support-v4:27.1.1'
-    testImplementation 'junit:junit:4.12'
-    androidTestImplementation 'com.android.support.test:runner:1.0.2'
-    androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
-    implementation 'com.google.android.gms:play-services-safetynet:16.0.0'
+    implementation 'androidx.appcompat:appcompat:1.2.0'
+    implementation 'androidx.constraintlayout:constraintlayout:2.0.0'
+    implementation "androidx.swiperefreshlayout:swiperefreshlayout:1.1.0"
+    implementation 'com.google.android.material:material:1.2.0'
+
+    testImplementation 'junit:junit:4.13'
+    androidTestImplementation 'androidx.test.ext:junit:1.1.1'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.2.0'
+    implementation 'com.google.android.gms:play-services-safetynet:17.0.0'
 
     // ButterKnife
     //noinspection GradleDependency
-    implementation 'com.jakewharton:butterknife:8.8.1'
-    annotationProcessor 'com.jakewharton:butterknife-compiler:8.8.1'
+    implementation 'com.jakewharton:butterknife:10.2.3'
+    annotationProcessor 'com.jakewharton:butterknife-compiler:10.2.3'
 
     //My depends
-    implementation 'com.squareup.okhttp3:okhttp:3.13.1'
+    implementation 'com.squareup.okhttp3:okhttp:4.8.1'
 
     //noinspection GradleDependency
-    implementation group: 'commons-io', name: 'commons-io', version: '2.6'
-    implementation 'commons-validator:commons-validator:1.6'
-    implementation 'com.felipecsl:gifimageview:2.2.0'
-    implementation 'joda-time:joda-time:2.10.1'
+    implementation group: 'commons-io', name: 'commons-io', version: '2.7'
+    implementation 'commons-validator:commons-validator:1.7'
+    implementation 'joda-time:joda-time:2.10.6'
     //noinspection GradleDependency
-    implementation 'com.github.bumptech.glide:glide:4.8.0'
+    implementation 'com.github.bumptech.glide:glide:4.11.0'
+    annotationProcessor 'com.github.bumptech.glide:compiler:4.11.0'
     //noinspection GradleDependency
-    implementation 'com.droidninja:filepicker:2.1.5'
-    implementation 'com.pes.materialcolorpicker:library:1.2.4'
-    implementation 'com.android.support:exifinterface:27.1.1'
+    implementation 'com.droidninja:filepicker:2.2.4'
+    implementation 'com.pes.materialcolorpicker:library:1.2.5'
+    implementation 'androidx.exifinterface:exifinterface:1.2.0'
 }
