@@ -2,8 +2,9 @@ package org.dreamexposure.startapped.activities;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
 import android.widget.Toast;
+
+import androidx.appcompat.app.AppCompatActivity;
 
 import org.dreamexposure.startapped.activities.auth.LoginActivity;
 import org.dreamexposure.startapped.async.TaskCallback;
