@@ -1,10 +1,11 @@
 package org.dreamexposure.startapped.activities.settings;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
 import android.widget.Switch;
 import android.widget.Toast;
+
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
 
 import org.dreamexposure.startapped.R;
 import org.dreamexposure.startapped.async.TaskCallback;
